+++ src/hmm/seq_show_files.c
@@ -25,7 +25,6 @@ int seq_show_files(
     );
 
     Hmm *hmm = 0;
-    HmmProb *hmmprob_object = 0;
 
     if (with_prob || gen_Qopt) {
         assert(hmm_filename != 0 && hmm_filename[0] != 0);
@@ -36,10 +35,6 @@ int seq_show_files(
             return 1;
         }
         printf("Loaded model for class: '%s'\n", hmm->className);
-
-        if (with_prob) {
-            hmmprob_object = hmmprob_create(hmm);
-        }
     }
 
     for (int i = 0; i < num_seq_filenames; ++i) {
@@ -62,12 +57,12 @@ int seq_show_files(
         }
 
         if (with_prob) {
-            prob_t prob = hmmprob_log_prob(hmmprob_object, seq, T);
+            prob_t prob = hmm_log_prob(hmm, seq, T);
             printf(" log(P) = %Le", prob);
         }
 
         if (gen_Qopt) {
-            int *Qopt = (int *) new_vector(T, sizeof(int));
+            int Qopt[T];
             prob_t probQopt = hmm_genQopt(hmm, seq, T, Qopt);
 
             printf("\n  N=%-4d Q* = «", hmm->N);
@@ -81,16 +76,10 @@ int seq_show_files(
             if (with_prob) {
                 printf(" log(P*) = %Le", probQopt);
             }
-
-            free(Qopt);
         }
         free(seq);
 
         printf("\n");
-    }
-
-    if (hmmprob_object) {
-        hmmprob_destroy(hmmprob_object);
     }
 
     return 0;